Refatoracao do codigo

+++ src/main/java/com/dev/loja/service/VendaService.java
@@ -8,7 +8,10 @@ import com.dev.loja.enums.PedidoStatus;
 import com.dev.loja.enums.ProdutoStatus;
 import com.dev.loja.exception.BadRequestException;
 import com.dev.loja.exception.EntityNotFoundException;
-import com.dev.loja.model.*;
+import com.dev.loja.model.ItemPedido;
+import com.dev.loja.model.Lancamento;
+import com.dev.loja.model.Pedido;
+import com.dev.loja.model.Produto;
 import com.dev.loja.repository.*;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
@@ -32,24 +35,33 @@ public class VendaService {
 
 
     public ResponseEntity<?> fecharPedido(@Valid CarrinhoDto carrinho, UserDetails userDetails) {
+
         if(carrinho.itens().isEmpty())
             throw new BadRequestException("Nenhum item foi adicionado ao carrinho");
 
-        var busca = userRepository.findUserByLogin(userDetails.getUsername());
-
-        if(busca.isEmpty())
-            throw new EntityNotFoundException("Cliente não encontrado");
-
-        if(busca.get().getEnderecos().isEmpty())
-            throw new BadRequestException("O cliente não tem nenhum endereço cadastrado");
-
         Pedido pedido = new Pedido();
-        pedido.setUser(busca.get());
+        pedido.setUser(null);
+        pedido.setCep(carrinho.cep());
+
+        if(userDetails != null){
+            var busca = userRepository.findUserByLogin(userDetails.getUsername());
+
+            if(busca.isEmpty())
+                throw new EntityNotFoundException("Cliente não encontrado");
+
+            if(busca.get().getEnderecos().isEmpty())
+                throw new BadRequestException("O cliente não tem nenhum endereço cadastrado");
+
+            pedido.setUser(busca.get());
+        }
+
         pedido.setFormaPagamento(carrinho.formaPagamento());
         pedido.setData(LocalDateTime.now());
         pedido.setFrete(Util.calculaFrete("17548700", carrinho.cep()));//implementar
 
-        vendaRepository.save(pedido);
+        if(userDetails != null) {
+            vendaRepository.save(pedido);
+        }
         return this.adicionarItens(pedido, carrinho.itens());
     }
 
@@ -63,7 +75,7 @@ public class VendaService {
 
         for(CarrinhoItem carrinhoItem : itensDuplicadosRemovidos){
             if(lancamentoRepository.consultarEstoque(carrinhoItem.produtoId) < carrinhoItem.quantidade){
-                vendaRepository.delete(pedido);
+                if(pedido.getNumero() !=null) vendaRepository.delete(pedido);
                 throw new BadRequestException("Estoque insuficiente para o produto de id "+
                         carrinhoItem.produtoId);
             }
@@ -77,17 +89,29 @@ public class VendaService {
                             .multiply(new BigDecimal(itemPedido.getQuantidade())));
             totalPedido = totalPedido.add(itemPedido.getSubtotal());//Acumula o subtotal
             itensPedido.add(itemPedido);
-            this.atualizarEstoque(carrinhoItem);
+            if(pedido.getNumero() !=null)
+                this.atualizarEstoque(carrinhoItem);
         }
-        pedido.setItens(itemPedidoRepository.saveAll(itensPedido));
+
+        if(pedido.getNumero() !=null)
+            pedido.setItens(itemPedidoRepository.saveAll(itensPedido));
+        else
+            pedido.setItens(itensPedido);
+
         pedido.setTotal(totalPedido);
         BigDecimal descontos = this.calculaDesconto(pedido);
         pedido.setDescontos(descontos);
         totalPedido = totalPedido.add(pedido.getFrete()).subtract(descontos);
         pedido.setTotal(totalPedido);
         pedido.setPedidoStatus(PedidoStatus.FECHADO);
-        var pedidoSalvo = vendaRepository.save(pedido);
-        return new ResponseEntity<>(new PedidoDtoSaida(pedidoSalvo), HttpStatus.CREATED);
+
+        if(pedido.getNumero() !=null){
+            var pedidoSalvo = vendaRepository.save(pedido);
+            return new ResponseEntity<>(new PedidoDtoSaida(pedidoSalvo), HttpStatus.CREATED);
+        }
+        return new ResponseEntity<>(new PedidoDtoSaida(pedido), HttpStatus.CREATED);
+
+
     }
 
     private Set<CarrinhoItem> somaRepetidos(List<CarrinhoItem> itens) {
@@ -117,6 +141,10 @@ public class VendaService {
                         .setScale(2, BigDecimal.ROUND_HALF_EVEN);}
             case PIX -> {
                 return new BigDecimal("0.10")
+                        .multiply(pedido.getTotal())
+                        .setScale(2, BigDecimal.ROUND_HALF_EVEN);}
+            case MERCADO_PAGO -> {
+                return new BigDecimal("0.03")
                         .multiply(pedido.getTotal())
                         .setScale(2, BigDecimal.ROUND_HALF_EVEN);}
             default -> {return BigDecimal.ZERO;}
